bug fix and search implement

+++ SysProfessor/FrmDisciplines.cs
@@ -43,39 +43,80 @@ namespace SysProfessor
             this.DgvDisciplines.Columns[2].AutoSizeMode = DataGridViewAutoSizeColumnMode.AllCells;
         }
 
-        private void Del()
-        {
-            string disciplineName;
-
-            int id = Convert.ToInt32(this.DgvDisciplines.CurrentRow.Cells["idmateria"].Value);
-
-            disciplineName = Convert.ToString(this.DgvDisciplines.CurrentRow.Cells["nome"].Value);
-
-            if (MessageBox.Show(
-                "Realmente Deseja Apagar a Matéria: " + disciplineName + " ?",
-                "Apagar Matéria?",
-                MessageBoxButtons.OKCancel,
-                MessageBoxIcon.Question) == DialogResult.OK)
-            {
-                string resp;
-
-                resp = Data.DeleteDiscipline(id);
-
-                Debug.WriteLine(resp);
-            }
-        }
-
         private void Search()
         {
             this.DgvDisciplines.DataSource = Data.SearchDiscipline(this.TxtSearch.Text);
         }
 
-        private void BtnView_Click(object sender, EventArgs e)
+        private void View()
         {
-            FormLoader.OpenChildForm(new FrmViewDiscipline(
+            if (DgvDisciplines.Rows.Count > 0)
+            {
+                //Debug.WriteLine("Tem dados");
+
+                FormLoader.OpenChildForm(new FrmViewDiscipline(
                 Convert.ToInt32(this.DgvDisciplines.CurrentRow.Cells["idmateria"].Value),
                 Convert.ToString(this.DgvDisciplines.CurrentRow.Cells["nome"].Value),
-                Convert.ToDecimal(this.DgvDisciplines.CurrentRow.Cells["media"].Value) ));
+                Convert.ToDecimal(this.DgvDisciplines.CurrentRow.Cells["media"].Value)));
+            }
+            else
+            {
+                Debug.WriteLine("não tem dados");
+            }
+        }
+
+        private void Del()
+        {
+            if (DgvDisciplines.Rows.Count > 0)
+            {
+                //Debug.WriteLine("Tem dados");
+
+                string disciplineName;
+
+                int id = Convert.ToInt32(this.DgvDisciplines.CurrentRow.Cells["idmateria"].Value);
+
+                disciplineName = Convert.ToString(this.DgvDisciplines.CurrentRow.Cells["nome"].Value);
+
+                if (MessageBox.Show(
+                    "Realmente Deseja Apagar a Matéria: " + disciplineName + " ?",
+                    "Apagar Matéria?",
+                    MessageBoxButtons.OKCancel,
+                    MessageBoxIcon.Question) == DialogResult.OK)
+                {
+                    string resp;
+
+                    resp = Data.DeleteDiscipline(id);
+
+                    Debug.WriteLine(resp);
+                }
+            }
+            else
+            {
+                Debug.WriteLine("não tem dados");
+            }
+        }
+
+        private void Add()
+        {
+            if (DgvDisciplines.Rows.Count > 0)
+            {
+                //Debug.WriteLine("Tem dados");
+
+                FormLoader.OpenChildForm(new FrmAddEditDiscipline(true,
+                Convert.ToInt32(this.DgvDisciplines.CurrentRow.Cells["idmateria"].Value),
+                Convert.ToString(this.DgvDisciplines.CurrentRow.Cells["nome"].Value),
+                Convert.ToString(this.DgvDisciplines.CurrentRow.Cells["media"].Value)
+                ));
+            }
+            else
+            {
+                Debug.WriteLine("não tem dados");
+            }
+        }
+
+        private void BtnView_Click(object sender, EventArgs e)
+        {
+            View();
         }
 
         private void BtnAdd_Click(object sender, EventArgs e)
@@ -85,11 +126,7 @@ namespace SysProfessor
 
         private void BtnEdit_Click(object sender, EventArgs e)
         {
-            FormLoader.OpenChildForm(new FrmAddEditDiscipline(true,
-                Convert.ToInt32(this.DgvDisciplines.CurrentRow.Cells["idmateria"].Value),
-                Convert.ToString(this.DgvDisciplines.CurrentRow.Cells["nome"].Value),
-                Convert.ToString(this.DgvDisciplines.CurrentRow.Cells["media"].Value)
-                ));
+            Add();
         }
 
         private void BtnDelete_Click(object sender, EventArgs e)
